requirements.txt
streamlit>=1.32.0
streamlit-js-eval>=1.0.0
PyMuPDF>=1.24.0
pandas>=2.2.0
thefuzz>=0.22.1
RapidFuzz>=3.10.0
Pillow>=10.0.0
